Created basic framework
App starts the project. GameMain has the main game loop. GameComponent displays everything to the screen. Phase is an enum for the current phase. Currently, only a black screen is shown. There is no way to start the game. Even if the game was started, the user has no control.